Decompose MCP client responsibilities
Separate transport, discovery, redaction, and result handling so lifecycle code stays focused while preserving the hardened limits and cleanup contracts.
